git-svn-id: http://192.168.20.11/sequoiadb/trunk@10251 6121f079-4a18-41d3-8c30-8870a322f2c2

+++ SequoiaDB/engine/dms/dmsInspect.cpp
@@ -117,7 +117,7 @@ namespace engine
             ++localErr ;
       }
 
-      if ( header->_storageUnitSize - dataOffset / pageSize != pageNum )
+      if ( header->_storageUnitSize - dataOffset / pageSize != (UINT32)pageNum )
       {
          len += ossSnprintf ( outBuf + len, outSize - len,
                               "Error: Page number is not same with value that "
